# Conflicts:
#	client.dat
#	src/com/pavogt/javaisland/data/Client.java
#	src/com/pavogt/javaisland/data/Product.java
#	src/com/pavogt/javaisland/screen/Admin.java

+++ src/com/pavogt/javaisland/LoginManager.java
@@ -78,7 +78,7 @@ public class LoginManager implements DataBaseListener {
         }
     }
 
-    public void register(String username, String password, String email, float balance) {
+    public void register(String username, String password, String email, int balance) {
         long hash = hashPassword(username, password);
         long uuid = db.getNextUuid();
         Client user = new Client(db, uuid, username, email, hash, balance, false);
